+++ CHANGELOG.md
@@ -9,6 +9,13 @@ genauso wie an [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Patch-Versionen nicht extra aufgeführt, sondern in der Minor-Version.
 
 
+## [Unreleased]
+
+### Changed
+
+- _fixed_: Vorsatz und Nachsatz im Datenpaket haben jetzt die gleiche Config wie das Datenpaket
+
+
 ## [6.1.0] - 2022-03-04
 
 ### Added